+++ ingenialink/ethernet/dictionary.py
@@ -4,8 +4,9 @@ from xml.etree import ElementTree
 
 import ingenialogger
 
-from ingenialink.dictionary import DictionaryV2, Interface
+from ingenialink.dictionary import DictionarySafetyModule, DictionaryV2, Interface
 from ingenialink.enums.register import RegAccess, RegCyclicType, RegDtype
+from ingenialink.ethercat.register import EthercatRegister
 from ingenialink.ethernet.register import EthernetRegister
 
 logger = ingenialogger.get_logger(__name__)
@@ -43,6 +44,14 @@ class EthernetDictionaryV2(DictionaryV2):
                 access=RegAccess.WO,
             ),
         ]
+
+    @cached_property
+    def _safety_registers(self) -> list[EthercatRegister]:
+        raise NotImplementedError("Safety registers are not implemented for this device.")
+
+    @cached_property
+    def _safety_modules(self) -> list[DictionarySafetyModule]:
+        raise NotImplementedError("Safety modules are not implemented for this device.")
 
     def _read_xdf_register(self, register: ElementTree.Element) -> Optional[EthernetRegister]:
         current_read_register = super()._read_xdf_register(register)