+++ ingenialink/servo.py
@@ -925,18 +925,11 @@ class Servo:
     def monitoring_channel_data(
         self,
         channel: int,
-        dtype: Optional[RegDtype] = None,  # noqa: ARG002
     ) -> list[float]:
         """Obtain processed monitoring data of a channel.
 
         Args:
             channel: Identity channel number.
-            dtype: Data type of the register to map.
-
-        Note:
-            The dtype argument is not necessary for this function, it
-            was added to maintain compatibility with IPB's implementation
-            of monitoring.
 
         Returns:
             Monitoring data.
@@ -1427,7 +1420,16 @@ class Servo:
             self._disturbance_write_data(chunk)
         self.disturbance_data = data
 
-    def _monitoring_read_data(self, **kwargs: Any) -> bytes:
+    def _is_monitoring_implemented(self) -> bool:
+        """Checks if monitoring is supported by the device.
+
+        Returns:
+            True if monitoring is implemented.
+
+        """
+        return not (self.MONITORING_DATA not in self.dictionary.registers(0))
+
+    def _monitoring_read_data(self) -> bytes:
         """Read monitoring data frame.
 
         Raises:
@@ -1437,24 +1439,36 @@ class Servo:
         Returns:
             monitoring data read.
         """
-        if self.MONITORING_DATA not in self.dictionary.registers(0):
+        if not self._is_monitoring_implemented():
             raise NotImplementedError("Monitoring is not supported by this device.")
-        if not isinstance(data := self.read(self.MONITORING_DATA, subnode=0, **kwargs), bytes):
+        if not isinstance(data := self.read(self.MONITORING_DATA, subnode=0), bytes):
             raise ValueError(
                 f"Error reading monitoring data. Expected type bytes, got {type(data)}"
             )
         return data
 
-    def _disturbance_write_data(self, data: bytes, **kwargs: Any) -> None:
+    def _is_disturbance_implemented(self) -> bool:
+        """Checks if disturbance is supported by the device.
+
+        Returns:
+            True if disturbance is implemented.
+
+        """
+        return not (self.DIST_DATA not in self.dictionary.registers(0))
+
+    def _disturbance_write_data(self, data: bytes) -> None:
         """Write disturbance data.
+
+        Args:
+            data: Data to be written.
 
         Raises:
             NotImplementedError: If disturbance is not supported by the device.
 
         """
-        if self.DIST_DATA not in self.dictionary.registers(0):
+        if not self._is_disturbance_implemented():
             raise NotImplementedError("Disturbance is not supported by this device.")
-        return self.write(self.DIST_DATA, subnode=0, data=data, **kwargs)
+        return self.write(self.DIST_DATA, subnode=0, data=data)
 
     @abstractmethod
     def _write_raw(self, reg: Register, data: bytes) -> None: