+++ tests/test_dictionary.py
@@ -22,6 +22,7 @@ PATH_RESOURCE = "./tests/resources/"
 PATH_TO_DICTIONARY = "./virtual_drive/resources/virtual_drive.xdf"
 
 
+@pytest.mark.no_connection
 @pytest.mark.parametrize(
     "dict_path, interface, fw_version, product_code, part_number, revision_number",
     [
@@ -71,6 +72,7 @@ def test_dictionary_description(
     )
 
 
+@pytest.mark.no_connection
 @pytest.mark.parametrize(
     "dict_path, interface, raises",
     [
@@ -134,6 +136,8 @@ def test_dictionary_v2_image_none(dictionary_class, dictionary_path):
         (f"{PATH_RESOURCE}ethercat/test_dict_ethercat.xdf", Interface.EoE, EthernetDictionaryV2),
         (f"{PATH_RESOURCE}test_dict_ecat_eoe_v3.0.xdf", Interface.ECAT, DictionaryV3),
         (f"{PATH_RESOURCE}test_dict_ecat_eoe_v3.0.xdf", Interface.EoE, DictionaryV3),
+        (f"{PATH_RESOURCE}test_dict_ecat_eoe_safe_v3.0.xdf", Interface.ECAT, DictionaryV3),
+        (f"{PATH_RESOURCE}test_dict_ecat_eoe_safe_v3.0.xdf", Interface.EoE, DictionaryV3),
     ],
 )
 def test_dictionary_factory(dict_path, interface, dict_class):